fix(web): 保持期間バッチの上限を有限にし、監査の失敗を別枠で数える
R2 head の失敗（監査が機能していない）を正常な競合の skippedRows と同じ箱に
入れていたため、検出が止まっていても info のままだった。auditErrors に分けて
cron の severity を warn 以上にする。

監査の抽出を ORDER BY RANDOM() から主キーの範囲走査へ変えた。前者は ready 行を
全件読んで並べ替えるため行数に比例して重くなる。ランダムな開始点（short_id と
同じ base62 12 文字）から LIMIT 件を読み、0 件なら先頭へ回り込む。

1 回の実行の subrequest が有限になるよう、期限切れ動画の SELECT と captures の
list ページ数にも上限を置き、打ち切りを sweepCapped に反映した。各フェーズの
最悪ケースの合計が Workers の上限（1000）を下回るよう pending の上限を 150 へ下げた。

retention.test.ts が 500 行上限を超えていたため、フェイクを helpers へ出し、
captures と pending 孤児のテストを別ファイルへ分けた（期待値は変えていない）。

検証: bun test（396 pass）/ bun run typecheck / check-architecture.py 9 件（baseline 10 件から減、新規なし）
Refs #66
