+++ web/src/lib/services/retention-audit.ts
@@ -6,7 +6,7 @@
  * 判定を誤ると再生できる動画を消してしまうため、回復は人の判断に残す。
  */
 
-import { movieKey } from '../contracts/r2key';
+import { generateShortId, movieKey } from '../contracts/r2key';
 
 /**
  * 1 回の実行で実体の有無を確かめる ready 行の上限。1 行につき head を 1 回使う。
@@ -34,8 +34,11 @@ export interface ReadyObjectAudit {
   checkedReadyRows: number;
   /** 実体が見つからなかった ready 行の数。0 以外は不変条件が破れた印。 */
   missingObjectRows: number;
-  /** R2 の一時失敗で判断できず、次回に委ねた行の数。 */
-  skippedRows: number;
+  /**
+   * R2 の head に失敗して判断できなかった行の数。監査そのものが機能していない印なので、
+   * 正常な競合によるスキップとは別に数えて cron のログで警告する。
+   */
+  auditErrors: number;
 }
 
 interface ShortIdRow {
@@ -48,22 +51,21 @@ interface ShortIdRow {
  * ready は commit が head で実体を確認してから付ける状態なので、実体の無い ready 行は
  * 削除の片落ち（services/movies.ts の deleteMovie）でしか生まれない。
  *
- * 毎回ランダムに選ぶのは、先頭から固定数を見ると同じ行だけを見続けて残りに永久に
- * 届かないため。1 回では取りこぼすが、毎時の実行を重ねれば行き渡る（カーソルのような
- * 状態をどこにも持たずに済む）。
+ * 抜き取りは「ランダムな開始点からの主キーの範囲走査」で行う。ORDER BY RANDOM() は
+ * ready 行を全件読んで並べ替えるため、行が増えるほど 1 回の実行が重くなる。short_id は
+ * 主キー（base62 12 文字）なので、同じ形式の値を毎回作って開始点にすれば、走査は
+ * LIMIT 件で止まりつつ実行のたびに違う範囲を見られる。
  */
 export async function auditReadyObjects(
   database: AuditDatabase,
-  bucket: AuditBucket
+  bucket: AuditBucket,
+  generateStartId: () => string = generateShortId
 ): Promise<ReadyObjectAudit> {
-  const { results } = await database
-    .prepare("SELECT short_id FROM movies WHERE status = 'ready' ORDER BY RANDOM() LIMIT ?")
-    .bind(MAX_OBJECT_CHECKS_PER_RUN)
-    .all<ShortIdRow>();
+  const results = await sampleReadyRows(database, generateStartId());
 
   let checkedReadyRows = 0;
   let missingObjectRows = 0;
-  let skippedRows = 0;
+  let auditErrors = 0;
 
   for (const row of results) {
     let object: { size: number } | null;
@@ -71,7 +73,7 @@ export async function auditReadyObjects(
       object = await bucket.head(movieKey(row.short_id));
     } catch {
       // R2 の一時失敗を「実体が無い」と読むと誤検知になる。判断は次回に委ねる。
-      skippedRows += 1;
+      auditErrors += 1;
       continue;
     }
 
@@ -87,5 +89,30 @@ export async function auditReadyObjects(
     if (remaining.results.length > 0) missingObjectRows += 1;
   }
 
-  return { checkedReadyRows, missingObjectRows, skippedRows };
+  return { checkedReadyRows, missingObjectRows, auditErrors };
+}
+
+/**
+ * 開始点以降の ready 行を LIMIT 件だけ読む。0 件なら先頭から取り直す（ラップ）。
+ *
+ * 開始点が既存の short_id より後ろに寄ると空になるため、回り込みが無いと末尾付近を
+ * 引いた実行が毎回何も見ないまま終わる。
+ */
+async function sampleReadyRows(database: AuditDatabase, startId: string): Promise<ShortIdRow[]> {
+  const page = await selectReadyRowsFrom(database, startId);
+  if (page.length > 0) return page;
+  return selectReadyRowsFrom(database, '');
+}
+
+async function selectReadyRowsFrom(
+  database: AuditDatabase,
+  startId: string
+): Promise<ShortIdRow[]> {
+  const { results } = await database
+    .prepare(
+      "SELECT short_id FROM movies WHERE status = 'ready' AND short_id >= ? ORDER BY short_id LIMIT ?"
+    )
+    .bind(startId, MAX_OBJECT_CHECKS_PER_RUN)
+    .all<ShortIdRow>();
+  return results;
 }
